book1-image1(please ignore last commit)

+++ raytracer/src/main.rs
@@ -154,7 +154,7 @@ fn main() {
 
     let aspect_ratio: f64 = 16.0 / 9.0;
     let image_width: u32 = 400;
-    let image_height: u32 = image_width / (aspect_ratio as u32);
+    let image_height: u32 = (image_width as f64 / aspect_ratio) as u32;
 
     let viewport_height = 2.0;
     let viewport_width = aspect_ratio * viewport_height;
